code/requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
blessings==1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler
dgl==0.6.1
dgl-cu111==0.7.2
dill
fairscale==0.4.5
filelock==3.6.0
fonttools==4.25.0
fugashi==1.1.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.44.0
huggingface-hub==0.4.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.11.1
ipadic==1.0.0
joblib
kiwisolver
Markdown==3.3.6
matplotlib
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multiprocess
munkres==1.1.4
networkx==2.6.3
numpy==1.22.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
olefile
omegaconf==2.1.1
packaging
pandas==1.4.2
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil
pytz==2022.1
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn
scipy==1.8.0
seaborn==0.11.2
sip==4.19.13
six
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl
tokenizers==0.11.5
torch
torchvision
tornado
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
